Ufagit add -A! Acabei de fazer a separação do lote e validade em uma tabela a parte

+++ src/model/entradamodel.cpp
@@ -1,4 +1,5 @@
 ﻿#include "entradamodel.h"
+#include "exception/exception.h"
 
 EntradaModel::EntradaModel()
 {
@@ -72,6 +73,21 @@ void EntradaModel::salvaListEntradaDeMaterial(const EntradaDeMaterialList& list)
 
 void EntradaModel::excluirEntrada(string id)
 {
+    EntradaDeMaterialList emList = entradaDeMaterialModel.getListEntradaDeMaterial(id);
+    if(saidaDeMaterialModel.existeSaidaApartirDasEntradas(emList))
+        throw Exception("Inpossível remover entradas que ja existem saida");
+
     dao->executeUpdate("DELETE FROM entrada_de_material WHERE entrada_id=" + id);
+
+    for(EntradaDeMaterialPtr& em: *emList){
+        MateiralPtr material = em->getMaterial();
+        material->setQuantidade(em->getQuantidade());
+        materialModel.decrementaQuantidadeMaterial(material);
+        LotePtr lote = em->getLote();
+        lote->setQuantidade(em->getQuantidade());
+        loteModel.decrementaLote(*lote);
+        if(lote->getQuantidade()<1)
+            loteModel.removeLote(*lote);
+    }
     dao->executeUpdate("DELETE FROM entrada WHERE id=" + id);
 }
